Henting av historisk pensjon skal brukes som en egen komponent
